+++ .editorconfig
@@ -1,11 +1,17 @@
+# editorconfig.org
 root = true
 
 [*]
 indent_style = space
-indent_size = 2
+indent_size = 4
+end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
 
 [*.md]
 trim_trailing_whitespace = false
+
+[{package,bower}.json]
+indent_style = space
+indent_size = 2
