+++ .gitignore
@@ -2,7 +2,10 @@ node_modules
 coverage
 .idea
 *.iml
+*.ipr
+*.iws
 atlassian-ide-plugin.xml
 /.project
 test/temp/
-npm-debug.log
+*debug.log*
+.vscode
