+++ generators/app/templates/test/_test-app.js
@@ -7,25 +7,25 @@ const deps = [
     [helpers.createDummyGenerator(), 'jhipster:modules']
 ];
 
-describe('JHipster generator <%= moduleName %>', function () {
-    describe('simple test', function () {
-        before(function (done) {
+describe('JHipster generator <%= moduleName %>', () => {
+    describe('simple test', () => {
+        beforeEach((done) => {
             helpers
-            .run(path.join( __dirname, '../generators/app'))
-            .inTmpDir(function (dir) {
-                fse.copySync(path.join(__dirname, '../test/templates/default'), dir)
-            })
-            .withOptions({
-                testmode: true
-            })
-            .withPrompts({
-                message: 'simple message to say hello'
-            })
-            .withGenerators(deps)
-            .on('end', done);
+                .run(path.join( __dirname, '../generators/app'))
+                .inTmpDir(function (dir) {
+                    fse.copySync(path.join(__dirname, '../test/templates/default'), dir)
+                })
+                .withOptions({
+                    testmode: true
+                })
+                .withPrompts({
+                    message: 'simple message to say hello'
+                })
+                .withGenerators(deps)
+                .on('end', done);
         });
 
-        it('generate dummy.txt file', function () {
+        it('generate dummy.txt file', () => {
             assert.file([
                 'dummy.txt'
             ]);
